Move files to the rebase and test/rebase folders
