+++ packages/dds/tree/src/rebase/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from "./rebaser";
+export * from "./verifyChangeRebaser";
